pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-my_env"
version = "1.0.0"
description = "MedCodeRL - Medical Coding & Billing Compliance environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    # Environment-specific dependencies
    "openai>=1.50.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
server = "my_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["my_env", "my_env.server"]
package-dir = { "my_env" = ".", "my_env.server" = "server" }

[tool.setuptools.package-data]
"my_env" = ["tasks/*.json"]